Dilemma over optimization algorithms, addition of loadable AIF values, problematic addition of constraints
Much has been added in this update, most summarized in the title above.
The Nelder Mead optimization algorithm creates many outliers, and so
would benefit from constraints. This lead me to the Coybal constrained
optimization algorithm, and while this creates maps that attach well to
real-world features and fits curves in seemingly every case (with not
outliers!), it catastrophically fails on the phantom, and seems to be
sensitive to both initial conditions and constraints. It's hard to say,
however, that it is worse than Nelder-Mead on real world data because of
the outliers. Maybe there is another constrained optimization algorithm
(that would need to be coded from scratch?) that can solve these
problems. Coybal is also a bit slower than Nelder-Mead. Other insights:
situations in which ktrans explodes and ve heads to 0 seem to mostly
represent the tofts model attempting to fit to noise. This combination
creates a flat curve with one tight peak, often minimizing "error."
Also, the constrained algorithms performance on non-Parker AIFs and
T1Maps is still up in the air - it's hard to know where to put the
constraints. But I suppose our argument is that one should always use
the Parker AIF..