Long day of testing optimization algorithms; untested batch mode; minor edits; version testing
Most edits pertain to the optimization loop, which is now a bit messier.
The Cobyal algorithm provides the best optimization visually, but is
slow and does not work on the phantom in some cases (particularly for
low ve + high ktrans, low ktrans high ve, i.e. the most difficult parts
to fit). For now, I'm running the Nelder-Mead algorithm on all points
because it is fast and accurate but prone to blowing up, and then
running the Cobyal algorithm on the points where it blows up, and then
finally masking the points that continue to escape. This is pretty
ad-hoc/unjustifiable, so a better solution is needed soon. It doesn't
even get rid of all the outliers.. The Cobyal algorithm feels the best -
if only it could work on the phantom and was faster. A bounded
Nelder-Mead might also work -- maybe one can be hand-coded from Matlab's
fminsearch model. All other optimization models in scipy perform worse
than these two. Also added a batch mode, and am working on making this
code usable as a package. Multiprocessing seems to not like being
packaged, which is very unfortunate and hopefully not a problem in
actual practice.. Next up: config files and a friendlier UI.